HDF5: update to new module structure
